+++ src/app/modules/slormancer/constants/common.ts
@@ -17,7 +17,7 @@ export const MAXIMUM_ANCESTRAL_LEGACY_POINTS = 7;
 
 export const DELIGHTED_VALUE = 12;
 
-export const GAME_VERSION = '0.3.014';
+export const GAME_VERSION = '0.3.015';
 
 export const POISON_DAMAGE_PERCENT = 200;
 export const RAVENOUS_DAGGER_DAMAGE_PERCENT = 100;